# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RedSecTools API",
    "version": "2.0.0",
    "description": "Repository-only OpenAPI documentation for RedSecTools. This file is not served by the live application. It documents the current HTTP API, including auth requirements for main app, guest-link, extension, and admin routes. RedSecThreat now includes user-scoped keyword, tag, and alert ownership on top of shared feed ingestion, plus admin-managed feed, template, and notification policy routes. Survey and Wiki include their active create/manage/respond/results/documentation route groups alongside the existing Paste, Share, Vault, Chat, Calendar, Homepage, Extension, and Admin surfaces."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_session",
        "description": "Signed user session cookie for the main web app."
      },
      "GuestSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_guest",
        "description": "Signed guest cookie created from /guest/{token}. Tool-scoped for paste or share."
      },
      "ExtensionBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque token",
        "description": "Bearer token returned by /api/ext/auth/login or /api/ext/auth/login/mfa."
      },
      "AdminCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "redsec_admin",
        "description": "Signed admin session cookie for /admin/api routes."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "retryAfter": { "type": "integer" },
          "restartLogin": { "type": "boolean" }
        },
        "required": ["error"]
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" }
        },
        "required": ["success"]
      },
...
      "ThreatFeed": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "feedType": { "type": "string", "enum": ["rss", "website", "api", "onion"] },
          "enabled": { "type": "boolean" },
          "isDefault": { "type": "boolean" },
          "fetchInterval": { "type": "integer", "description": "Polling interval in seconds (minimum 60)" },
          "feedMetadata": { "type": "string", "description": "JSON string with feed-type-specific configuration" },
          "lastFetchedAt": { "type": ["string", "null"], "format": "date-time" },
          "lastHash": { "type": ["string", "null"] },
          "lastError": { "type": ["string", "null"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ThreatFeedRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "url": { "type": "string" },
          "feedType": { "type": "string", "enum": ["rss", "website", "api", "onion"] },
          "enabled": { "type": "boolean", "default": true },
          "isDefault": { "type": "boolean", "default": false },
          "fetchInterval": { "type": "integer", "description": "Seconds between fetches (minimum 60, default 3600)" },
          "feedMetadata": { "type": "object", "description": "Feed-type-specific config (auth, headers, field mapping for API feeds; proxy config for onion)" },
          "tagIds": { "type": "array", "items": { "type": "string" } },
          "keywordIds": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["name", "url", "feedType"]
      },
...
      "ThreatNotificationConfig": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "channelType": { "type": "string", "enum": ["webhook", "email", "discord"] },
          "destination": { "type": "string" },
          "enabled": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ThreatNotificationRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "channelType": { "type": "string", "enum": ["webhook", "email", "discord"] },
          "destination": { "type": "string" },
          "enabled": { "type": "boolean", "default": true }
        },
        "required": ["name", "channelType", "destination"]
      },
      "ThreatApiTemplate": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "configuration": { "type": "object", "description": "Endpoint, method, headers, auth, field_mapping" },
          "enabled": { "type": "boolean" },
          "isSystem": { "type": "boolean", "description": "System templates cannot be deleted" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ThreatApiTemplateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "configuration": { "type": "object" },
          "enabled": { "type": "boolean", "default": true }
        },
        "required": ["name"]
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or expired",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/threat/feeds": {
      "get": {
        "tags": ["Threat"],
        "summary": "List threat feeds",
        "description": "Returns all configured feed sources. Query param ?enabled=true filters to enabled only. Requires threat.view.",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [
          { "name": "enabled", "in": "query", "required": false, "schema": { "type": "string", "enum": ["true", "1"] }, "description": "Filter to enabled feeds only" }
        ],
        "responses": {
          "200": { "description": "Feed list", "content": { "application/json": { "schema": { "type": "object", "properties": { "feeds": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatFeed" } } } } } } },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Threat"],
        "summary": "Create a threat feed",
        "description": "Public user sessions cannot create feeds here. Feed source management lives in the admin panel under /admin/api/threat/feeds.",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreatFeedRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Feed created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "feed": { "$ref": "#/components/schemas/ThreatFeed" } } } } } },
          "400": { "description": "Validation error (missing name/url, invalid feed type)" },
          "403": { "description": "Feed source changes are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/feeds/{id}": {
      "get": {
        "tags": ["Threat"],
        "summary": "Get a single threat feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Feed details", "content": { "application/json": { "schema": { "type": "object", "properties": { "feed": { "$ref": "#/components/schemas/ThreatFeed" } } } } } },
          "404": { "description": "Feed not found" },
          "403": { "description": "Access denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["Threat"],
        "summary": "Update a threat feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThreatFeedRequest" } } }
        },
        "responses": {
          "200": { "description": "Feed updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "feed": { "$ref": "#/components/schemas/ThreatFeed" } } } } } },
          "400": { "description": "Validation error" },
          "404": { "description": "Feed not found" },
          "403": { "description": "Management denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Threat"],
        "summary": "Delete a threat feed",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Feed deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Success" } } } },
          "400": { "description": "Invalid feed ID" },
          "403": { "description": "Management denied" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/threat/notifications": {
      "get": {
        "tags": ["Threat"],
        "summary": "List admin notification configs",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": { "description": "Notification configs", "content": { "application/json": { "schema": { "type": "object", "properties": { "notifications": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatNotificationConfig" } } } } } } },
          "403": { "description": "Threat notification policy is only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Threat"],
        "summary": "Create a notification config",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreatNotificationRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Notification created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "notification": { "$ref": "#/components/schemas/ThreatNotificationConfig" } } } } } },
          "400": { "description": "Validation error" },
          "403": { "description": "Threat notification policy is only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/notifications/{id}": {
      "put": {
        "tags": ["Threat"],
        "summary": "Update a notification config",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "destination": { "type": "string" },
                  "enabled": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Notification updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "notification": { "$ref": "#/components/schemas/ThreatNotificationConfig" } } } } } },
          "404": { "description": "Notification not found" },
          "403": { "description": "Threat notification policy is only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Threat"],
        "summary": "Delete a notification config",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Notification deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Success" } } } },
          "403": { "description": "Threat notification policy is only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/threat/templates": {
      "get": {
        "tags": ["Threat"],
        "summary": "List API feed templates",
        "security": [{ "UserSessionCookie": [] }],
        "responses": {
          "200": { "description": "Template list", "content": { "application/json": { "schema": { "type": "object", "properties": { "templates": { "type": "array", "items": { "$ref": "#/components/schemas/ThreatApiTemplate" } } } } } } },
          "403": { "description": "Threat API templates are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Threat"],
        "summary": "Create an API feed template",
        "security": [{ "UserSessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ThreatApiTemplateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Template created", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "template": { "$ref": "#/components/schemas/ThreatApiTemplate" } } } } } },
          "400": { "description": "Validation error" },
          "403": { "description": "Threat API templates are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/templates/{id}": {
      "put": {
        "tags": ["Threat"],
        "summary": "Update an API feed template",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ThreatApiTemplateRequest" } }
          }
        },
        "responses": {
          "200": { "description": "Template updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "template": { "$ref": "#/components/schemas/ThreatApiTemplate" } } } } } },
          "404": { "description": "Template not found" },
          "403": { "description": "Threat API templates are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Threat"],
        "summary": "Delete an API feed template",
        "description": "Public user sessions cannot delete templates here. Template management lives in the admin panel.",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Template deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Success" } } } },
          "403": { "description": "Threat API templates are only available in the admin panel" },
          "404": { "description": "Template not found" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/threat/templates/{id}/test": {
      "post": {
        "tags": ["Threat"],
        "summary": "Test an API feed template",
        "description": "Public user sessions cannot test templates here. Template testing lives in the admin panel.",
        "security": [{ "UserSessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Template test result", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "result": { "type": "object" } } } } } },
          "404": { "description": "Template not found" },
          "500": { "description": "Template test failed" },
          "403": { "description": "Threat API templates are only available in the admin panel" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  }
}